Revert "Feat activity adapter"
