+++ README.md
@@ -23,6 +23,7 @@ Init the audit object:
 require_once __DIR__ . '/../../vendor/autoload.php';
 
 use PDO;
+use PDO;
 use Utopia\Audit\Audit;
 use Utopia\Cache\Cache;
 use Utopia\Cache\Adapter\None as NoCache;
@@ -42,7 +43,7 @@ $pdo = new PDO("mysql:host={$dbHost};port={$dbPort};charset=utf8mb4", $dbUser, $
     PDO::ATTR_EMULATE_PREPARES => true,
     PDO::ATTR_STRINGIFY_FETCHES => true,
 ]);
-
+        
 $cache = new Cache(new NoCache());
 
 $database = new Database(new MySQL($pdo),$cache);
