+++ composer.json
@@ -21,9 +21,8 @@
         "utopia-php/database": "0.60.*"
     },
     "require-dev": {
-        "ext-pdo": "*",
         "phpunit/phpunit": "9.*",
-        "phpstan/phpstan": "2.*",
+        "phpstan/phpstan": "1.*",
         "laravel/pint": "1.*"
     },
     "config": {
