+++ docker-compose.yml
@@ -1,3 +1,5 @@
+version: '3'
+
 services:
   mariadb:
     image: mariadb:10.11
