--- src/Audit/Adapter/Audit.php
@@ -1,134 +1,0 @@
-<?php
-
-namespace Utopia\Audit\Adapter;
-
-use Utopia\Audit\Adapter;
-use Utopia\Audit\Log;
-use Utopia\Database\Database;
-
-class Audit extends Adapter
-{
-    public function getCollection(): string
-    {
-        return 'audit';
-    }
-
-    public function getAttributes(): array
-    {
-        return [
-            [
-                '$id' => 'userId',
-                'type' => Database::VAR_STRING,
-                'size' => Database::LENGTH_KEY,
-                'required' => true,
-                'signed' => true,
-                'array' => false,
-                'filters' => [],
-            ], [
-                '$id' => 'event',
-                'type' => Database::VAR_STRING,
-                'size' => 255,
-                'required' => true,
-                'signed' => true,
-                'array' => false,
-                'filters' => [],
-            ], [
-                '$id' => 'resource',
-                'type' => Database::VAR_STRING,
-                'size' => 255,
-                'required' => false,
-                'signed' => true,
-                'array' => false,
-                'filters' => [],
-            ], [
-                '$id' => 'userAgent',
-                'type' => Database::VAR_STRING,
-                'size' => 65534,
-                'required' => true,
-                'signed' => true,
-                'array' => false,
-                'filters' => [],
-            ], [
-                '$id' => 'ip',
-                'type' => Database::VAR_STRING,
-                'size' => 45,
-                'required' => true,
-                'signed' => true,
-                'array' => false,
-                'filters' => [],
-            ], [
-                '$id' => 'location',
-                'type' => Database::VAR_STRING,
-                'size' => 45,
-                'required' => false,
-                'signed' => true,
-                'array' => false,
-                'filters' => [],
-            ], [
-                '$id' => 'time',
-                'type' => Database::VAR_DATETIME,
-                'format' => '',
-                'size' => 0,
-                'signed' => true,
-                'required' => false,
-                'array' => false,
-                'filters' => ['datetime'],
-            ], [
-                '$id' => 'data',
-                'type' => Database::VAR_STRING,
-                'size' => 16777216,
-                'required' => false,
-                'signed' => true,
-                'array' => false,
-                'filters' => ['json'],
-            ],
-        ];
-    }
-
-    public function getIndexes(): array
-    {
-        return [
-            [
-                '$id' => 'index2',
-                'type' => Database::INDEX_KEY,
-                'attributes' => ['event'],
-                'lengths' => [],
-                'orders' => [],
-            ], [
-                '$id' => 'index4',
-                'type' => Database::INDEX_KEY,
-                'attributes' => ['userId', 'event'],
-                'lengths' => [],
-                'orders' => [],
-            ], [
-                '$id' => 'index5',
-                'type' => Database::INDEX_KEY,
-                'attributes' => ['resource', 'event'],
-                'lengths' => [],
-                'orders' => [],
-            ], [
-                '$id' => 'index-time',
-                'type' => Database::INDEX_KEY,
-                'attributes' => ['time'],
-                'lengths' => [],
-                'orders' => [Database::ORDER_DESC],
-            ],
-        ];
-    }
-
-    public function filter(Log $log): Log
-    {
-        unset($log['hostname']);
-        unset($log['projectId']);
-        unset($log['projectInternalId']);
-        unset($log['resourceId']);
-        unset($log['resourceInternalId']);
-        unset($log['resourceParent']);
-        unset($log['resourceType']);
-        unset($log['teamId']);
-        unset($log['teamInternalId']);
-        unset($log['userInternalId']);
-        unset($log['userType']);
-        return $log;
-    }
-}
